+++ flutter-hms-awareness/android/src/main/java/com/huawei/hms/flutter/awareness/constants/Channel.java
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2022. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -21,5 +21,5 @@ public interface Channel {
     String BARRIER_CHANNEL = "com.huawei.hms.flutter.awareness/barrier";
     String BARRIER_LISTENER_CHANNEL = "com.huawei.hms.flutter.awareness/barrierListener";
     String UTILS_CHANNEL = "com.huawei.hms.flutter.awareness/utils";
-    String PERMISSION_CHANNEL = "com.huawei.hms.flutter.awareness/permissions";
+
 }