+++ flutter-hms-awareness/lib/huawei_awareness.dart
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2022. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -17,7 +17,6 @@
 library huawei_awareness;
 
 import 'dart:convert';
-import 'dart:typed_data';
 
 import 'package:flutter/services.dart';