+++ flutter-hms-awareness/lib/src/constants/channel.dart
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2022. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -25,6 +25,4 @@ abstract class _Channel {
       'com.huawei.hms.flutter.awareness/barrierListener';
   static const String awarenessUtilsChannel =
       'com.huawei.hms.flutter.awareness/utils';
-  static const String awarenessPermissionCahnnel =
-      'com.huawei.hms.flutter.awareness/permissions';
 }