+++ flutter-hms-awareness/lib/src/constants/method.dart
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2022. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -47,20 +47,8 @@ abstract class _Method {
   static const String queryBarriers = 'queryBarriers';
   static const String deleteBarrier = 'deleteBarrier';
   static const String updateBarrier = 'updateBarriers';
-  static const String updateBarrierWithAutoRemove =
-      'UpdateBarriersWithAutoRemove';
 
   //Awareness Utils Client Methods
   static const String enableLogger = 'enableLogger';
   static const String disableLogger = 'disableLogger';
-  static const String hasLocationPermission = 'hasLocationPermission';
-  static const String hasBackgroundLocationPermission =
-      'hasBackgroundLocationPermission';
-  static const String hasActivityRecognitionPermission =
-      'hasActivityRecognitionPermission';
-  static const String requestLocationPermission = 'requestLocationPermission';
-  static const String requestBackgroundLocationPermission =
-      'requestBackgroundLocationPermission';
-  static const String requestActivityRecognitionPermission =
-      'requestActivityRecognitionPermission';
 }