+++ flutter-hms-awareness/lib/src/constants/param.dart
@@ -1,5 +1,5 @@
 /*
-    Copyright 2020-2022. Huawei Technologies Co., Ltd. All rights reserved.
+    Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License")
     you may not use this file except in compliance with the License.
@@ -17,12 +17,12 @@
 part of huawei_awareness;
 
 abstract class _Param {
-  //Capture Client Parameters
+  // Capture Client Parameters
   static const String countryCode = 'countryCode';
   static const String futureTimestamp = 'futureTimestamp';
   static const String packageName = 'packageName';
 
-  //Common Parameters
+  // Common Parameters
   static const String barrierEventType = 'barrierEventType';
   static const String barrierType = 'barrierType';
   static const String barrierLabel = 'barrierLabel';
@@ -52,7 +52,7 @@ abstract class _Param {
   static const String ssid = 'ssid';
   static const String status = 'status';
 
-  //AmbientLight Parameters
+  // AmbientLight Parameters
   static const String ambientLightBarrierReceiverAction =
       'AMBIENT_LIGHT_BARRIER_RECEIVER_ACTION';
   static const String ambientLightAboveBarrier =
@@ -64,34 +64,34 @@ abstract class _Param {
   static const String minLightIntensity = 'minLightIntensity';
   static const String maxLightIntensity = 'maxLightIntensity';
 
-  //Barrier Parameters
+  // Barrier Parameters
   static const String barrierStatus = 'barrierStatus';
 
-  //Barrier Delete Request Parameters
+  // Barrier Delete Request Parameters
   static const String deleteType = 'deleteType';
   static const String barrierKey = 'barrierKey';
   static const String deleteAll = 'deleteAll';
   static const String withLabel = 'withLabel';
 
-  //Barrier Query Request Parameters
+  // Barrier Query Request Parameters
   static const String queryTypeAll = 'queryTypeAll';
   static const String queryTypeKey = 'queryTypeKey';
   static const String queryType = 'queryType';
   static const String barrierKeys = 'barrierKeys';
 
-  //Barrier Status
+  // Barrier Status
   static const String lastBarrierUpdateTime = 'lastBarrierUpdateTime';
   static const String lastStatus = 'lastStatus';
   static const String presentStatus = 'presentStatus';
 
-  //Beacon Barrier
+  // Beacon Barrier
   static const String beaconBarrierReceiverAction =
       'BEACON_BARRIER_RECEIVER_ACTION';
   static const String beaconDiscoverBarrier = 'BEACON_DISCOVER_BARRIER_LABEL';
   static const String beaconMissedBarrier = 'BEACON_MISSED_BARRIER_LABEL';
   static const String beaconKeepBarrier = 'BEACON_KEEP_BARRIER_LABEL';
 
-  //Behavior Barrier
+  // Behavior Barrier
   static const String behaviorBarrierReceiverAction =
       'BEHAVIOR_BARRIER_RECEIVER_ACTION';
   static const String behaviorKeepingBarrier = 'BEHAVIOR_KEEPING_BARRIER_LABEL';
@@ -100,7 +100,7 @@ abstract class _Param {
   static const String behaviorEndingBarrier = 'BEHAVIOR_ENDING_BARRIER_LABEL';
   static const String behaviorTypes = 'behaviorTypes';
 
-  //Bluetooth Barrier
+  // Bluetooth Barrier
   static const String bluetoothBarrierReceiverAction =
       'BLUETOOTH_BARRIER_RECEIVER_ACTION';
   static const String bluetoothKeepBarrier = 'BLUETOOTH_KEEP_BARRIER_LABEL';
@@ -109,7 +109,7 @@ abstract class _Param {
   static const String bluetoothDisconnectingBarrier =
       'BLUETOOTH_DISCONNECTING_BARRIER_LABEL';
 
-  //Combination Barrier
+  // Combination Barrier
   static const String combinedBarrierReceiverAction =
       'COMBINED_BARRIER_RECEIVER_ACTION';
   static const String combinedNot = 'COMBINED_NOT';
@@ -117,7 +117,7 @@ abstract class _Param {
   static const String combinedOr = 'COMBINED_OR';
   static const String barrier = 'barrier';
 
-  //Headset Barrier
+  // Headset Barrier
   static const String headsetBarrierReceiverAction =
       'HEADSET_BARRIER_RECEIVER_ACTION';
   static const String headsetKeepingBarrier = 'HEADSET_KEEPING_BARRIER_LABEL';
@@ -126,7 +126,7 @@ abstract class _Param {
   static const String headsetDisconnectingBarrier =
       'HEADSET_DISCONNECTING_BARRIER_LABEL';
 
-  //Location Barrier
+  // Location Barrier
   static const String locationBarrierReceiverAction =
       'LOCATION_BARRIER_RECEIVER_ACTION';
   static const String locationEnterBarrier = 'LOCATION_ENTER_BARRIER_LABEL';
@@ -135,7 +135,7 @@ abstract class _Param {
   static const String radius = 'radius';
   static const String timeOfDuration = 'timeOfDuration';
 
-  //Screen Barrier
+  // Screen Barrier
   static const String screenBarrierReceiverAction =
       'SCREEN_BARRIER_RECEIVER_ACTION';
   static const String screenKeepingBarrier = 'SCREEN_KEEPING_BARRIER_LABEL';
@@ -143,7 +143,7 @@ abstract class _Param {
   static const String screenOffBarrier = 'SCREEN_OFF_BARRIER_LABEL';
   static const String screenUnlockBarrier = 'SCREEN_UNLOCK_BARRIER_LABEL';
 
-  //Time Barrier
+  // Time Barrier
   static const String timeBarrierReceiverAction =
       'TIME_BARRIER_RECEIVER_ACTION';
   static const String timeInSunriseOrSunsetPeriodBarrier =
@@ -168,7 +168,7 @@ abstract class _Param {
   static const String stopTimeOfSpecifiedDay = 'stopTimeOfSpecifiedDay';
   static const String inTimeCategory = 'inTimeCategory';
 
-  //WiFi Barrier
+  // WiFi Barrier
   static const String wifiBarrierReceiverAction =
       'WIFI_BARRIER_RECEIVER_ACTION';
   static const String wifiKeepingBarrier = 'WIFI_KEEPING_BARRIER_LABEL';
@@ -179,7 +179,7 @@ abstract class _Param {
   static const String wifiDisabling = 'WIFI_DISABLING_BARRIER_LABEL';
   static const String wifiStatus = 'wifiStatus';
 
-  //Aqi Parameters
+  // Aqi Parameters
   static const String co = 'co';
   static const String no2 = 'no2';
   static const String o3 = 'o3';
@@ -187,15 +187,15 @@ abstract class _Param {
   static const String pm25 = 'pm25';
   static const String so2 = 'so2';
 
-  //City Parameters
+  // City Parameters
   static const String cityCode = 'cityCode';
   static const String name = 'name';
   static const String provinceName = 'provinceName';
 
-  //Daily Live Info Parameters
+  // Daily Live Info Parameters
   static const String level = 'level';
 
-  //Daily Weather Parameters
+  // Daily Weather Parameters
   static const String maxTempC = 'maxTempC';
   static const String maxTempF = 'maxTempF';
   static const String minTempC = 'minTempC';
@@ -208,17 +208,17 @@ abstract class _Param {
   static const String sunRise = 'sunRise';
   static const String sunSet = 'sunSet';
 
-  //Hourly Weather Parameters
+  // Hourly Weather Parameters
   static const String isDayNight = 'isDayNight';
   static const String rainProbability = 'rainProbability';
   static const String tempC = 'tempC';
   static const String tempF = 'tempF';
 
-  //Live Info Parameters
+  // Live Info Parameters
   static const String code = 'code';
   static const String dailyLiveInfo = 'dailyLiveInfo';
 
-  //Situation Parameters
+  // Situation Parameters
   static const String humidity = 'humidity';
   static const String pressure = 'pressure';
   static const String realFeelC = 'realFeelC';
@@ -228,34 +228,34 @@ abstract class _Param {
   static const String updateTime = 'updateTime';
   static const String uvIndex = 'uvIndex';
 
-  //Weather Situation Parameters
+  // Weather Situation Parameters
   static const String situation = 'situation';
 
-  //Application Response Parameters
+  // Application Response Parameters
   static const String applicationStatus = 'applicationStatus';
 
-  //Beacon Response Parameters
+  // Beacon Response Parameters
   static const String beacons = 'beacons';
 
-  //Behavior Response Parameters
+  // Behavior Response Parameters
   static const String elapsedRealtimeMillis = 'elapsedRealtimeMillis';
   static const String mostLikelyBehavior = 'mostLikelyBehavior';
   static const String probableBehavior = 'probableBehavior';
 
-  //Capability Response Parameters
+  // Capability Response Parameters
   static const String deviceSupportCapabilities = 'deviceSupportCapabilities';
 
-  //Dark Mode Response Parameters
+  // Dark Mode Response Parameters
   static const String isDarkModeOn = 'isDarkModeOn';
 
-  //Detected Behavior Parameters
+  // Detected Behavior Parameters
   static const String confidence = 'confidence';
   static const String type = 'type';
 
-  //Light Intensity Response Parameters
+  // Light Intensity Response Parameters
   static const String lightIntensity = 'lightIntensity';
 
-  //Location Response Parameters
+  // Location Response Parameters
   static const String altitude = 'altitude';
   static const String speed = 'speed';
   static const String bearing = 'bearing';
@@ -266,30 +266,30 @@ abstract class _Param {
       'speedAccuracyMetersPerSecond';
   static const String fromMockProvider = 'fromMockProvider';
 
-  //Time Categories Parameters
+  // Time Categories Parameters
   static const String timeCategories = 'timeCategories';
 
-  //Weather Position Parameters
+  // Weather Position Parameters
   static const String locale = 'locale';
   static const String country = 'country';
   static const String province = 'province';
   static const String district = 'district';
   static const String county = 'county';
 
-  //Weather Response Parameters
+  // Weather Response Parameters
   static const String dailyWeather = 'dailyWeather';
   static const String hourlyWeather = 'hourlyWeather';
   static const String liveInfo = 'liveInfo';
   static const String aqi = 'aqi';
   static const String weatherSituation = 'weatherSituation';
 
-  //Beacon Filter Parameters
+  // Beacon Filter Parameters
   static const String matchByNameType = 'matchByNameType';
   static const String matchByBeaconContent = 'matchByBeaconContent';
   static const String matchByBeaconId = 'matchByBeaconId';
   static const String filterType = 'filterType';
 
-  //Update Barrier Parameters
+  // Update Barrier Parameters
   static const String autoRemove = 'autoRemove';
   static const String request = 'request';
 }