+++ flutter-hms-awareness/CHANGELOG.md
@@ -1,8 +1,13 @@
+## 3.8.0+300
+
+- **Breaking Change:** With this release, permission-related methods have been removed. You are expected to handle the required permissions on your own. You can learn more about the required permissions from our [official documentations](https://developer.huawei.com/consumer/en/doc/HMS-Plugin-Guides/before-starting-0000001073950576-V1?ha_source=hms1).
+- API 33 support has been added.
+
 ## 3.1.0+302
 
 - [Breaking Change] Updated to null safety.
 - [Breaking Change] Modified the names of all constant variables as lowerCamelCase.
-- Modified the internal structure of the plugin. Please use import 'package:huawei_awareness/huawei_awareness.dart'; not to get any errors.
+- Modified the internal structure of the plugin. Please use `import 'package:huawei_awareness/huawei_awareness.dart';` not to get any errors.
 
 ## 1.0.7+304